Completely remove CumulusCI 1 (Ant based) from repo:
- Deleted the `build` and `ci` directories.  Deleting `build` deleted
  the ant-salesforce.jar that was making repo 17MB bigger
- Deleted `cumulusci.tasks.ant.AntTask` which can now be more
  configurably replaced with `cumulusci.tasks.command.Command` or
`cumulusci.tasks.command.SalesforceCommand'
- Deleted the CumulusCI 1.0 CLI previously under cumulusci/cli and moved
  cumulusci/newcli to cumulusci/cli
