--- ci/cloud/README.md
@@ -1,1 +1,0 @@
-Experimental integration scripts for cloud based CI systems
