--- ci/cloud/semaphore/setup.sh
@@ -1,5 +1,0 @@
-# Install python libraries
-pip install --upgrade simple-salesforce
-pip install --upgrade selenium
-pip install --upgrade requests
-pip install --upgrade PyGithub==1.25.1
