+++ cumulusci/cli/logger.py
